+++ test/dummy/test/support/oauth_dummy_helpers.rb
@@ -91,12 +91,15 @@ module OauthDummyHelpers
     [root_recording, folder_recording, result.value]
   end
 
-  def create_oauth_client(name: "Demo App", confidential: false, redirect_uris: ["http://127.0.0.1/callback"], api: "public")
+  def create_oauth_client(name: "Demo App", confidential: false, redirect_uris: ["http://127.0.0.1/callback"], api: "public", use_central_relay: false, allowed_return_patterns: [], exact_return_urls: [])
     attrs = {
       name: name,
       confidential: confidential,
       redirect_uris: redirect_uris,
-      api_key: api.to_s
+      api_key: api.to_s,
+      use_central_relay: use_central_relay,
+      allowed_return_patterns: allowed_return_patterns,
+      exact_return_urls: exact_return_urls
     }
     secret_token = nil
     if confidential